requirements.txt
Django==1.6.3
Fabric==1.7.0
boto==2.6.0
django-social-auth==0.7.10
django-storages==1.1.5
httplib2==0.7.7
oauth2==1.5.211
python-openid==2.2.5
requests==1.1.0
wsgiref==0.1.2
gunicorn==0.14.6
dj-database-url==0.2.0
psycopg2==2.4.5
South==0.8.4
Pillow==2.3.1
dj-static==0.0.5
static==0.4
django-lockdown==1.0
django-pipeline==1.3.19
honcho==0.5.0
django-mail-templated==0.2.0

celery==3.1.7
django-celery==3.1.1
kombu==3.0.8
anyjson==0.3.3
amqp==1.3.3
billiard==3.3.0.13
pytz==2013.9
jsmin==2.0.9

supervisor==3.0
django-kombu==0.9.4

supervisor==3.0

yattag==0.9.4
